start breaking keyframe management code out into its own class so it can be used by independent drivers
